requirements.txt
pyxel
python-socketio
eventlet
requests
websocket-client

